Update configuration handling for config.yml using Configurate
This required creating FileKeys in the common module, while also creating instances of Fusion.

The options from the config.yml for the most part are "static" in the way that they update on /ce reload, or populate on server startup. That way we are not constantly checking config files directly on every single event, especially with some events being called hundreds of times.

new Gson() is now static in Methods, we can re-use the object.

I also moved more logic from listeners to avoid some other circular loading, they are now in the Methods class which will eventually be turned back into a static class. The utils directory/classes will be re-visited as well.